Module Tezos_alpha_test_helpers.Assert

val error : loc:string -> ('a, Tezos_error_monad.TzCore.error list) Stdlib.result -> (Tezos_error_monad.TzCore.error -> bool) -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val proto_error : loc:string -> ('a, Tezos_error_monad.TzCore.error list) Stdlib.result -> (Tezos_protocol_alpha.Protocol.Environment.Error_monad.error -> bool) -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val equal : loc:string -> ('a -> 'a -> bool) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> 'a -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val not_equal : loc:string -> ('a -> 'a -> bool) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> 'a -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

module Int32 : sig ... end

module Int64 : sig ... end

val equal_int : loc:string -> int -> int -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val equal_int32 : loc:string -> int32 -> int32 -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val not_equal_int : loc:string -> int -> int -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val leq_int : loc:string -> int -> int -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val equal_int64 : loc:string -> int64 -> int64 -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val not_equal_int64 : loc:string -> int64 -> int64 -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val equal_bool : loc:string -> bool -> bool -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val not_equal_bool : loc:string -> bool -> bool -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val equal_tez : loc:string -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val not_equal_tez : loc:string -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val equal_pkh : loc:string -> Tezos_crypto.Signature.Public_key_hash.t -> Tezos_crypto.Signature.Public_key_hash.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val not_equal_pkh : loc:string -> Tezos_crypto.Signature.Public_key_hash.t -> Tezos_crypto.Signature.Public_key_hash.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val balance_is : loc:string -> Tezos_alpha_test_helpers__Context.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> ?kind:Context.Contract.balance_kind -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

balance_is b c amount checks that the current balance of contract c is amount. Default balance type is Main, pass ~kind with Deposit, Fees or Rewards for the others.

val balance_was_operated : operand:('a -> 'b -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.t Tezos_protocol_alpha.Protocol.Environment.Error_monad.tzresult) -> loc:string -> Tezos_alpha_test_helpers__Context.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> ?kind:Context.Contract.balance_kind -> 'a -> 'b -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

balance_was_operated ~operand b c old_balance amount checks that the current balance of contract c is operand old_balance amount and returns the current balance. Default balance type is Main, pass ~kind with Deposit, Fees or Rewards for the others.

val balance_was_credited : loc:string -> Tezos_alpha_test_helpers__Context.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> ?kind:Context.Contract.balance_kind -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.tez -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.tez -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val balance_was_debited : loc:string -> Tezos_alpha_test_helpers__Context.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> ?kind:Context.Contract.balance_kind -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.tez -> Tezos_protocol_alpha.Protocol.Alpha_context.Tez.tez -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val print_balances : Tezos_alpha_test_helpers__Context.t -> Tezos_protocol_alpha.Protocol.Alpha_context.Contract.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t