Module Tezos_alpha_test_helpers__Contract_helpers

val init : unit -> (Tezos_alpha_test_helpers.Block.t * Tezos_raw_protocol_alpha__Alpha_context.public_key_hash * Tezos_raw_protocol_alpha.Alpha_context.Contract.t * Tezos_raw_protocol_alpha.Alpha_context.Contract.t, Tezos_base__TzPervasives.tztrace) Stdlib.result Lwt.t

Initializes 2 addresses to do only operations plus one that will be used to bake.

val toplevel_from_string : string -> Tezos_raw_protocol_alpha.Alpha_context.Script.expr

Parses a Michelson contract from string.

val expression_from_string : string -> Tezos_raw_protocol_alpha.Alpha_context.Script.expr

Parses a Michelson expression from string, useful for call parameters.

val originate_contract : string -> string -> Tezos_raw_protocol_alpha.Alpha_context.Contract.contract -> Tezos_alpha_test_helpers.Block.t -> Tezos_raw_protocol_alpha.Alpha_context.public_key_hash -> (Tezos_raw_protocol_alpha.Alpha_context.Contract.contract * Tezos_alpha_test_helpers.Block.t, Tezos_base__TzPervasives.tztrace) Stdlib.result Lwt.t

Returns a block in which the contract is originated.