Module Block.Forge

val contents : ?proof_of_work_nonce:Stdlib.Bytes.t -> ?priority:int -> ?seed_nonce_hash:Tezos_protocol_alpha.Protocol.Nonce_hash.t -> ?liquidity_baking_escape_vote:bool -> unit -> AC.Block_header.contents

type header

val forge_header : ?policy:baker_policy -> ?timestamp:AC.Timestamp.time -> ?operations:AC.Operation.packed list -> ?liquidity_baking_escape_vote:bool -> t -> (header, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Forges a correct header following the policy. The header can then be modified and applied with apply.

val set_seed_nonce_hash : Tezos_protocol_alpha.Protocol.Nonce_hash.t option -> header -> header

Sets uniquely seed_nonce_hash of a header

val set_baker : AC.public_key_hash -> header -> header

Sets the baker that will sign the header to an arbitrary pkh

val sign_header : header -> (AC.Block_header.block_header, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Signs the header with the key of the baker configured in the header. The header can no longer be modified, only applied.