Module Tezos_alpha_test_helpers.Context

type t =

| B of Block.t

| I of Incremental.t

val branch : t -> Tezos_protocol_alpha.Protocol.Environment.Block_hash.t

val get_level : t -> Tezos_raw_protocol_alpha.Alpha_context.Raw_level.t Tezos_base__TzPervasives.tzresult

val get_endorsers : t -> Tezos_raw_protocol_alpha.Alpha_services.Delegate.Endorsing_rights.t list Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_endorser : t -> (Tezos_raw_protocol_alpha.Alpha_context.public_key_hash * int list) Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_voting_power : t -> Tezos_raw_protocol_alpha.Alpha_context.public_key_hash -> int32 Tezos_protocol_alpha.Protocol.Environment.Error_monad.shell_tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_total_voting_power : t -> int32 Tezos_protocol_alpha.Protocol.Environment.Error_monad.shell_tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_bakers : t -> Tezos_raw_protocol_alpha.Alpha_context.public_key_hash list Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_seed_nonce_hash : t -> Tezos_raw_protocol_alpha.Nonce_hash.t Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_seed : t -> Tezos_raw_protocol_alpha.Alpha_context.Seed.seed Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

Returns the seed of the cycle to which the block belongs to.

val get_constants : t -> Tezos_raw_protocol_alpha.Alpha_context.Constants.t Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

Returns all the constants of the protocol

val get_minimal_valid_time : t -> priority:int -> endorsing_power:int -> Tezos_protocol_alpha.Protocol.Environment.Time.t Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_baking_reward : t -> priority:int -> endorsing_power:int -> Tezos_raw_protocol_alpha.Alpha_context.Tez.t Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_endorsing_reward : t -> priority:int -> endorsing_power:int -> Tezos_raw_protocol_alpha.Alpha_context.Tez.t Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_liquidity_baking_subsidy : t -> Tezos_raw_protocol_alpha.Alpha_context.Tez.t Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

val get_liquidity_baking_cpmm_address : t -> Tezos_raw_protocol_alpha.Alpha_context.Contract.t Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

module Vote : sig ... end

module Contract : sig ... end

module Delegate : sig ... end

val init : ?⁠rng_state:Stdlib.Random.State.t -> ?⁠endorsers_per_block:int -> ?⁠with_commitments:bool -> ?⁠initial_balances:int64 list -> ?⁠initial_endorsers:int -> ?⁠min_proposal_quorum:int32 -> ?⁠time_between_blocks:Tezos_raw_protocol_alpha.Alpha_context.Period.t list -> ?⁠minimal_block_delay:Tezos_raw_protocol_alpha.Alpha_context.Period.t -> ?⁠delay_per_missing_endorsement:Tezos_raw_protocol_alpha.Alpha_context.Period.t -> ?⁠bootstrap_contracts:Tezos_raw_protocol_alpha.Alpha_context.Parameters.bootstrap_contract list -> ?⁠level:int32 -> int -> (Block.t * Tezos_raw_protocol_alpha.Alpha_context.Contract.t list) Tezos_base__TzPervasives.tzresult Tezos_protocol_alpha.Protocol.Environment.Lwt.t

init n : returns an initial block with n initialized accounts and the associated implicit contracts