Module Tezos_alpha_test_helpers.Op

val endorsement : ?⁠delegate:Tezos_raw_protocol_alpha.Alpha_context.public_key_hash -> ?⁠level:Tezos_raw_protocol_alpha.Alpha_context.Raw_level.t -> Context.t -> ?⁠signing_context:Context.t -> unit -> Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement Tezos_raw_protocol_alpha.Alpha_context.Operation.t Tezos_base__TzPervasives.tzresult Lwt.t

val endorsement_with_slot : ?⁠delegate:(Tezos_raw_protocol_alpha.Alpha_context.public_key_hash * int list) -> ?⁠level:Tezos_raw_protocol_alpha.Alpha_context.Raw_level.t -> Context.t -> ?⁠signing_context:Context.t -> unit -> Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement_with_slot Tezos_raw_protocol_alpha.Alpha_context.Operation.t Tezos_base__TzPervasives.tzresult Lwt.t

val miss_signed_endorsement : ?⁠level:Tezos_raw_protocol_alpha.Alpha_context.Raw_level.t -> Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement Tezos_raw_protocol_alpha.Alpha_context.Operation.t Tezos_base__TzPervasives.tzresult Lwt.t

val transaction : ?⁠counter:Z.t -> ?⁠fee:Tezos_raw_protocol_alpha.Alpha_context.Tez.tez -> ?⁠gas_limit:Tezos_raw_protocol_alpha.Alpha_context.Gas.Arith.integral -> ?⁠storage_limit:Z.t -> ?⁠parameters:Tezos_raw_protocol_alpha.Alpha_context.Script.lazy_expr -> ?⁠entrypoint:string -> Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Contract.t -> Tezos_raw_protocol_alpha.Alpha_context.Contract.t -> Tezos_raw_protocol_alpha.Alpha_context.Tez.t -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed Tezos_base__TzPervasives.tzresult Lwt.t

val delegation : ?⁠fee:Tezos_raw_protocol_alpha.Alpha_context.Tez.tez -> Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Contract.t -> Tezos_raw_protocol_alpha.Alpha_context.public_key_hash option -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed Tezos_base__TzPervasives.tzresult Lwt.t

val revelation : ?⁠fee:Tezos_raw_protocol_alpha.Alpha_context.Tez.tez -> Context.t -> Tezos_raw_protocol_alpha.Alpha_context.public_key -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed Tezos_base__TzPervasives.tzresult Lwt.t

val failing_noop : Context.t -> Tezos_raw_protocol_alpha.Alpha_context.public_key_hash -> string -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed Tezos_base__TzPervasives.tzresult Lwt.t

val origination : ?⁠counter:Z.t -> ?⁠delegate:Tezos_raw_protocol_alpha.Alpha_context.public_key_hash -> script:Tezos_raw_protocol_alpha.Alpha_context.Script.t -> ?⁠preorigination:Tezos_raw_protocol_alpha.Alpha_context.Contract.contract option -> ?⁠public_key:Tezos_raw_protocol_alpha.Alpha_context.public_key -> ?⁠credit:Tezos_raw_protocol_alpha.Alpha_context.Tez.tez -> ?⁠fee:Tezos_raw_protocol_alpha.Alpha_context.Tez.tez -> ?⁠gas_limit:Tezos_raw_protocol_alpha.Alpha_context.Gas.Arith.integral -> ?⁠storage_limit:Z.t -> Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Contract.contract -> (Tezos_raw_protocol_alpha.Alpha_context.Operation.packed * Tezos_raw_protocol_alpha.Alpha_context.Contract.contract) Tezos_base__TzPervasives.tzresult Lwt.t

val originated_contract : Tezos_raw_protocol_alpha.Alpha_context.Operation.packed -> Tezos_raw_protocol_alpha.Alpha_context.Contract.contract

val double_endorsement : Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement Tezos_raw_protocol_alpha.Alpha_context.Operation.t -> Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement Tezos_raw_protocol_alpha.Alpha_context.Operation.t -> slot:int -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed

val double_baking : Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Block_header.block_header -> Tezos_raw_protocol_alpha.Alpha_context.Block_header.block_header -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed

val activation : Context.t -> Tezos_crypto.Signature.Public_key_hash.t -> Tezos_raw_protocol_alpha.Blinded_public_key_hash.activation_code -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed Tezos_base__TzPervasives.tzresult Lwt.t

val combine_operations : ?⁠public_key:Tezos_raw_protocol_alpha.Alpha_context.public_key -> ?⁠counter:Tezos_raw_protocol_alpha.Alpha_context.counter -> ?⁠spurious_operation:Tezos_raw_protocol_alpha.Alpha_context.packed_operation -> source:Tezos_raw_protocol_alpha.Alpha_context.Contract.t -> Context.t -> Tezos_raw_protocol_alpha.Alpha_context.packed_operation list -> Tezos_raw_protocol_alpha.Alpha_context.packed_operation Tezos_base__TzPervasives.tzresult Lwt.t

val seed_nonce_revelation : Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Raw_level.t -> Tezos_raw_protocol_alpha.Alpha_context.Nonce.t -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed

Reveals a seed_nonce that was previously committed at a certain level

val proposals : Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Contract.t -> Tezos_crypto.Protocol_hash.t list -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed Tezos_base__TzPervasives.tzresult Lwt.t

Propose a list of protocol hashes during the approval voting

val ballot : Context.t -> Tezos_raw_protocol_alpha.Alpha_context.Contract.t -> Tezos_crypto.Protocol_hash.t -> Tezos_raw_protocol_alpha.Alpha_context.Vote.ballot -> Tezos_raw_protocol_alpha.Alpha_context.Operation.packed Tezos_base__TzPervasives.tzresult Lwt.t

Cast a vote yay, nay or pass

val dummy_script : Tezos_raw_protocol_alpha.Alpha_context.Script.t

val dummy_script_cost : Test_tez.Tez.t